requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
fastapi-mcp==0.3.0
langchain>=0.3.15
langchain-openai>=0.2.14
langchain-mcp-tools==0.1.0
pygithub==2.5.0
docker==7.1.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.6.0

